[Cleanup] Revert program_cache.hpp surface-area reduction
